Roll of dice simulation
The roll of a dice is a typical example of random experiment because of the uncertainity of the result and the fact there is not dependence among the observations. This enable the apply several theorem which helps tho describe a random variable like the Law of Large Number or  Central Limit theorem.